requirements.txt
# Core Libraries (Original)
streamlit>=1.41.0
gspread==5.10.0
oauth2client==4.1.3
pandas
numpy>=2.0
fpdf2

# Google Auth Dependencies (Original)
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# NEW: Template Manager Support
google-api-python-client==2.108.0

# Utilities (Original)
python-dateutil==2.8.2
rapidfuzz==3.14.3
openpyxl>=3.1.0
